+++ ja2lib/Tactical/SoldierControl.c
@@ -5323,7 +5323,7 @@ UINT8 SoldierTakeDamage(struct SOLDIERTYPE *pSoldier, INT8 bHeight, INT16 sLifeD
     case ENEMY_TEAM:
       // if we're in the wilderness this always counts
       if (IsSectorEnemyControlled(gWorldSectorX, gWorldSectorY) ||
-          SectorInfo[GetSectorID8(gWorldSectorX, gWorldSectorY)]
+          SectorInfo[GetSectorID8((u8)gWorldSectorX, (u8)gWorldSectorY)]
                   .ubTraversability[THROUGH_STRATEGIC_MOVE] != TOWN) {
         // update current day of activity!
         UpdateLastDayOfPlayerActivity((UINT16)GetWorldDay());
